Refactor and relocate schema validation tests for `dateTimeReceived`, `disposal/recovery codes`, and `means of transport` to align with test structure cleanup.
